Fix Fivemat initialization with Cucumber 2
Cucumber 2.x has the concept of legacy formatters and modern formatters.
It checks formatters for the arity of the #initialize method to
determine whether the formatter is legacy or modern. Fivemat performs a
bit of trickery with a module .new method, meaning there is nothing to
check the arity on.

This hacks around this bit of neepery by checking if Cucumber 2.x is
loaded and overriding the #instance_method method to return the instance
method of Fivemat::Cucumber.